mcp-management/requirements.txt
# MCP Server Dependencies for LearnLine Educational CMS
# Install with: pip install -r requirements.txt

# Core MCP Protocol (install from GitHub until PyPI release)
# Temporary: We'll implement a lightweight MCP server without the official SDK
# The official SDK will be: mcp[cli]>=1.0.0 when available

# LLM Backends
google-genai>=1.5.0        # Gemini API (free tier: 1,000 requests/day)
anthropic>=0.40.0          # Claude API (already installed in main project)

# HTTP Client for calling FastAPI endpoints
httpx>=0.27.0

# WebSocket support
websockets>=12.0

# Optional: ASGI server for HTTP transport
uvicorn>=0.30.0

# JSON-RPC protocol implementation (for lightweight MCP)
jsonrpcserver>=5.0.0
jsonrpcclient>=4.0.0

# Async support
aiofiles>=24.1.0
